requirements.txt
selenium
requests
beautifulsoup4
schedule
python-telegram-bot==13.15
